Implement ByRef parameters and refactor plugin builds
Adds support for ByRef parameters in the CrossBasic VM, parser, and code generator, including a new ObjRef type and OP_GET_REF opcode. Refactors plugin build scripts for FileWatcher and SQLiteDatabase/SQLiteStatement to correct output names and linking. Cleans up legacy array assignment handling and improves type coercion for plugin callbacks. Also includes method overload support for classes and stricter numeric type checks in VM arithmetic operations. XGraphics/XPicture/XCanvas have been updated to be fully working on Windows. Work to add GTK and Cocoa is in-progress for cross-platform desktop UI controls.

+++ Plugins/PluginTemplate.cpp
@@ -45,7 +45,7 @@ extern "C" {
     } PluginEntry;
 
     // Sample exported function: adds two integers.
-    XPLUGIN_API int addtwonumbers(int a, int b) {
+    XPLUGIN_API double addtwonumbers(double a, double b) {
         return a + b;
     }
